Move writeback message_models to common.messages.writeback
Move lib dependency on os.environ to lambdas. Correct for previously introduced module name changes.
