+++ hasher-matcher-actioner/hmalib/common/tests/test_writebacker.py
@@ -6,7 +6,8 @@ import os
 
 from hmalib.lambdas.actions.writebacker import lambda_handler
 from hmalib.common.classification_models import WritebackTypes
-from hmalib.common.message_models import MatchMessage, WritebackMessage, BankedSignal
+from hmalib.common.messages.match import MatchMessage, BankedSignal
+from hmalib.common.messages.writeback import WritebackMessage
 from hmalib.common.configs.fetcher import ThreatExchangeConfig
 from hmalib.common import config as hmaconfig
