+++ hasher-matcher-actioner/hmalib/lambdas/pdq/pdq_matcher.py
@@ -14,7 +14,7 @@ from threatexchange.signal_type.pdq import PdqSignal
 
 from hmalib import metrics
 from hmalib.models import MatchRecord
-from hmalib.common.message_models import BankedSignal, MatchMessage
+from hmalib.common.messages.match import BankedSignal, MatchMessage
 from hmalib.common.models.signal import PDQSignalMetadata
 from hmalib.common.logging import get_logger
 from hmalib.common.config import HMAConfig
